+++ core/camerafixture/sceneFeaturePoint.h
@@ -186,7 +186,7 @@ public:
         this->type = type;
     }
 
-    Vector3dd triangulate(bool use__ = false, uint32_t mask = ~uint32_t(0));
+    Vector3dd triangulate(bool use__ = false, uint64_t mask = ~uint64_t(0));
 
 
     /** Observation related block */
